+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/networking/RangeUpgradeWirelessTransmitterRangeModifier.java
@@ -1,4 +1,4 @@
-package com.refinedmods.refinedstorage.common.wirelesstransmitter;
+package com.refinedmods.refinedstorage.common.networking;
 
 import com.refinedmods.refinedstorage.common.Platform;
 import com.refinedmods.refinedstorage.common.api.upgrade.UpgradeState;