+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/wirelesstransmitter/package-info.java
@@ -1,7 +1,0 @@
-@ParametersAreNonnullByDefault
-@FieldsAndMethodsAreNonnullByDefault
-package com.refinedmods.refinedstorage.common.wirelesstransmitter;
-
-import com.refinedmods.refinedstorage.api.core.FieldsAndMethodsAreNonnullByDefault;
-
-import javax.annotation.ParametersAreNonnullByDefault;